changed SPHoptions from int to uint64_t, added if statements that disable the gravity queueing

+++ basetype.h
@@ -174,7 +174,7 @@ typedef struct {
     double dFlopDoubleCPU;
     double dFlopDoubleGPU;
 
-    int SPHoptions;
+    uint64_t SPHoptions;
     } workParticle;
 
 /*
